requirements.txt
streamlit>=1.36
python-dotenv>=1.0.1
pypdf>=4.2.0
openai>=1.40.0
requests>=2.32.3